refactor: 大幅调整项目结构
