+++ src/www/app/components/README.md
@@ -1,0 +1,3 @@
+# Components
+
+This is the **Components** page of the documentation website.
