+++ src/www/app/components/shadcn/select/select-basic.ts
@@ -1,4 +1,4 @@
-import { Div, h } from '@/utils/dom';
+import { Div, h } from '@/www/utils/dom';
 import { DocCode } from '@/www/components/doc-component';
 import '@/shadcn';
 import { HighlightRule } from '@/www/components/doc-component';
