+++ src/www/app/docs/README.md
@@ -1,0 +1,3 @@
+# Docs
+
+This is the **Docs** page of the documentation website.
