--- src/www/app/shadcn/tab/index.ts
@@ -1,1 +1,0 @@
-import './tab';
