+++ src/www/components/READEME.md
@@ -1,0 +1,4 @@
+# Website Components
+
+These components are only used in the Prototype-UI documentation website.
+
