+++ src/www/components/doc-component/doc-anchor.ts
@@ -1,4 +1,4 @@
-import { A, Div, Li, P, Ul } from '@/utils/dom';
+import { A, Div, Li, P, Ul } from '@/www/utils/dom';
 
 class DocAnchor extends HTMLElement {
   connectedCallback() {
