+++ src/www/components/doc-component/doc-component.ts
@@ -1,5 +1,5 @@
 import { ContextProvider } from '@/common';
-import { A, Div, h, Main, P, PrototypeTabContent, Span } from '@/utils/dom';
+import { A, Div, h, Main, P, PrototypeTabContent, Span } from '@/www/utils/dom';
 import { Doc, DocContext } from './interface';
 
 export default abstract class DocComponent extends ContextProvider<DocContext> {
