+++ src/www/components/website-nav/website-nav.ts
@@ -1,6 +1,6 @@
 import '@/prototype/tab';
 import './github-button';
-import { Div, h, Nav, PrototypeTabTrigger, Span } from '@/utils/dom';
+import { Div, h, Nav, PrototypeTabTrigger, Span } from '@/www/utils/dom';
 
 export default class WebsiteNav extends HTMLElement {
   private _navItemsCls =
